Conflicts:
	ambry-api/src/main/java/com.github.ambry/network/Receive.java
	ambry-network/src/main/java/com.github.ambry.network/SocketServer.java
	ambry-network/src/main/java/com.github.ambry.network/SocketServerInputSet.java

+++ ambry-api/src/main/java/com.github.ambry/network/Receive.java
@@ -3,6 +3,7 @@ package com.github.ambry.network;
 import java.io.IOException;
 import java.nio.channels.ReadableByteChannel;
 
+
 /**
  * Used to receive data from the network channel. Any implementation of this interface
  * can be used to read data from the network
@@ -10,15 +11,16 @@ import java.nio.channels.ReadableByteChannel;
 public interface Receive {
   /**
    * Indicates if the read has been completed
-   * @return  true if read is complete, else false
+   * @return true if read is complete, else false
    */
   boolean isReadComplete();
 
   /**
    * Reads some bytes from the provided channel
    * @param channel The channel to read from
-   * @return No of bytes read. Returns -1 if EOS is reached
+   * @return Number of bytes read. Returns -1 if EOS is reached
    * @throws IOException
    */
-  int readFrom(ReadableByteChannel channel) throws IOException;
+
+  long readFrom(ReadableByteChannel channel) throws IOException;
 }
